Refine GN::Str class: Now the class holds one and only one pointer. So it can be safely used in printf function without having to convert to raw pointer.

+++ garnet3d/main/src/priv/inc/garnet/gfx/fatModel.h
@@ -116,8 +116,8 @@ namespace GN { namespace gfx
         StrA   material; // name of the material.
         uint32 basevtx;
         uint32 numvtx;
-        size_t startidx;
-        size_t numidx;
+        uint32 startidx;
+        uint32 numidx;
     };
 
     struct FatMesh
